Delete ClientTweaks absolete code.
Thanks to  
Contributor Casimir255 for all the code work here!
https://github.com/Casimir255/Essentials/commit/a23d4dbd49f2f476499c8078e5c2e4bc2d44c3de

So here we have deleted all client tweaks, that no longer used, and added checks to clean GPS marks of other players and last death positions of other players from sent data.
Also all the toolbars and research of all other players.

This will reduce the amount of DATA client need to get and proccess during login to server.
We have 3000 user accounts, this should really help!

* Also i have deleted no longer needed code for GameTags, KEEN fixed the issue.
We never used this.

* I have added check for Ranks system init and patch.
if not enabled, no point to patch the chat code and other stuff.