Econ Module can work with offlines now (#213)
* Fix: It possible KeyAlreadyExistException workarounded

It apparently is possible to have 2 Players with Same Steam ID but different Serial ID. Cause unknown.

For both of them the same Identity ID is looked up because that ignores the Serial ID. And therefore I can just replace the value.

Could auso check with if(!containsKey) but since the result will be identical I didn't bother.

* Fix: Credits are Number-Formatted in all eco commands

* Fix: Players can nolonger pay themselves

Players were able to duplicate money by paying themselves.
If Player A has 10,000 credits
And Pays Player A 1,000 credits
He ended up with 11,000 credits.

But there was a warning on server side that the client was out of sync. And if you do anything else, like for example !econ set !econ take or !econ reset it crashed said client.

* Fix: Added missing .DisplayName when Paying player

* Feat: !econ check can now also check offline players

I dont exactly know why that limitation of online only was in place. set, take, give, check should work for offlines too.

Since the others are not important enough (for me at least) I am not doing that with this commit.

* Fix: Responses now inform about the possability that the player is offline

These kind of commands should work with IMyIdentity instead of IMyPlayer especially since the "*" parameter would work for ALL Players (not NPCs)

but since I have no intention to fix it at this point in time. The better message will have to do.

* Feat: Econ Module can now also work with offline players

* Fix: If more credits are taken than the account has, credits are set to 0. And for transfers proper feedback is provided

* Fix: Wrong Sync Client Server fworks correctly now

Who would have guessed ChangeBalanceBroadcastClient does NOT broadcast to the client. ChangeBalance does that automatically.

The other method is... probably what is called IN the client after receiving the event.

* Fix: added forgotten formatting

* Fix: More Formatting